LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/access-control-preflight-request-invalid-status-501-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/XMLHttpRequest/resources/access-control-preflight-request-invalid-status.py?code=501. Preflight response is not successful

PASS Request with 501 preflight response 
